Rename frag prints module.
